+++ deepchecks/tabular/checks/data_integrity/is_single_value.py
