+++ tests/tabular/checks/integrity/data_duplicate_test.py
@@ -16,7 +16,7 @@ from hamcrest import (assert_that, calling, close_to, equal_to, has_items,
 
 from deepchecks.core import ConditionCategory
 from deepchecks.core.errors import DatasetValidationError, DeepchecksValueError
-from deepchecks.tabular.checks.integrity.data_duplicates import DataDuplicates
+from deepchecks.tabular.checks.data_integrity.data_duplicates import DataDuplicates
 from tests.base.utils import equal_condition_result
