+++ tests/tabular/checks/integrity/special_chars_test.py
@@ -16,7 +16,7 @@ from hamcrest import (assert_that, calling, close_to, has_items, has_length,
 
 from deepchecks.core import ConditionCategory
 from deepchecks.core.errors import DeepchecksValueError
-from deepchecks.tabular.checks.integrity.special_chars import SpecialCharacters
+from deepchecks.tabular.checks.data_integrity.special_chars import SpecialCharacters
 from deepchecks.tabular.dataset import Dataset
 from tests.base.utils import equal_condition_result
